+++ CHANGELOG.md
@@ -1,9 +1,20 @@
+## v1.2
+* releasing geo features, facets, writes, multi and monetize
+
+## v1.1.9
+* added geocode, geopulse and world-geographies
+* added facets
+* adding writes (submit, flag)
+* updated documents
+* added multi
+* added monetize
+
 ## v1.1
-* debug mode
-* raw read
+* added debug mode
+* added raw read
 
 ## v1.0.3
-* adding crosswalk(namespace_id, namespace)
+* added crosswalk(namespace_id, namespace)
 
 ## v1.0.2
 * lazy getting total_count of a query